move getLabwareNicknamesById and dependent selectors to new ui/labware
