+++ protocol-designer/src/labware-ingred/utils.js
@@ -3,9 +3,10 @@ import {humanizeLabwareType} from '@opentrons/components'
 import type {DisplayLabware} from './types'
 
 export const labwareToDisplayName = (
-  displayLabware: DisplayLabware,
+  displayLabware: ?DisplayLabware,
   labwareType: string
-) => (
-  displayLabware.nickname ||
-  `${humanizeLabwareType(labwareType)} (${displayLabware.disambiguationNumber})`
-)
+) => {
+  const disambiguationNumber = displayLabware ? displayLabware.disambiguationNumber : ''
+  return (displayLabware && displayLabware.nickname) ||
+  `${humanizeLabwareType(labwareType)} (${disambiguationNumber})`
+}
