+++ protocol-designer/src/ui/labware/index.js
@@ -1,0 +1,6 @@
+// @flow
+import * as selectors from './selectors'
+
+export {
+  selectors,
+}
